搜索模块导入数据库数据

+++ astc_parent/astc_service/astc_service_search/src/main/java/com/astc/search/dao/SkuMapper.java
@@ -4,6 +4,6 @@ package com.astc.search.dao;
 import com.astc.search.pojo.SkuInfo;
 import org.springframework.data.elasticsearch.repository.ElasticsearchRepository;
 
-public interface SkuMapper extends ElasticsearchRepository<SkuInfo, Integer> {
+public interface SkuMapper extends ElasticsearchRepository<SkuInfo, Long> {
 
 }
